working on create inventory page
