+++ src/app/admin/inventory/_components/inventory-list.tsx
@@ -31,7 +31,7 @@ export type InventoryWithRelations = Prisma.InventoryGetPayload<{
   };
 }>;
 
-export default function InventoryList({inventories}:{inventories:InventoryWithRelations}) {
+export default function InventoryList({inventories}:{inventories:InventoryWithRelations[]}) {
   return (
     <>
