+++ src/app/admin/inventory/new/page.tsx
@@ -1,34 +1,8 @@
-"use client";
-import Link from "next/link";
 import { ContentLayout } from "../../_components/content-layout";
 import DynamicBreadcrumb from "../../_components/dynamic-breadcrumb";
-import { Button } from "@/components/ui/button";
-import { ChevronLeft } from "lucide-react";
-import { Badge } from "@/components/ui/badge";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
+import { getProducts } from "../../products/actions";
+import CreateInventoryForm from "../_components/create-inventory-form";
 
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { productSchema } from "../../products/schema";
-
-import { useRef } from "react";
-import { Switch } from "@/components/ui/switch";
 
 const breadcrumbItems = [
   { label: "Dashboard", href: "/admin" },
@@ -40,243 +14,13 @@ const breadcrumbItems = [
   },
 ];
 
-const defaultValues = {
-  name: "",
-  description: "",
-  features: [],
-};
 
-export default function CreateInventoryPage() {
-  const form = useForm<z.infer<typeof productSchema>>({
-    resolver: zodResolver(productSchema),
-    defaultValues,
-    // mode: "all",
-  });
-
-  const formRef = useRef<HTMLFormElement>(null);
-
+export default async function CreateInventoryPage() {
+  const products  = await getProducts();
   return (
     <ContentLayout title="Add Inventory Item">
       <DynamicBreadcrumb items={breadcrumbItems} />
-      <Form {...form}>
-        <form>
-          <div className="flex items-center gap-4 mb-5 mt-7">
-            <Link href="/admin/products">
-              <Button variant="outline" size="icon" className="h-7 w-7">
-                <ChevronLeft className="h-4 w-4" />
-                <span className="sr-only">Back</span>
-              </Button>
-            </Link>
-
-            <h1 className="flex-1 shrink-0 whitespace-nowrap text-xl font-semibold tracking-tight sm:grow-0">
-              Pro Controller
-            </h1>
-            <Badge variant="outline" className="ml-auto sm:ml-0">
-              In stock
-            </Badge>
-            <div className="hidden items-center gap-2 md:ml-auto md:flex">
-              <Button variant="outline" size="sm">
-                Discard
-              </Button>
-              <Button
-                size="sm"
-                type="submit"
-                // disabled={!form.formState.isValid}
-              >
-                Save Product
-              </Button>
-            </div>
-          </div>
-
-          <div className="grid gap-4 md:grid-cols-[1fr_250px] lg:grid-cols-3 lg:gap-8">
-            <div className="grid auto-rows-max items-start gap-4 lg:col-span-2 lg:gap-8">
-              <Card x-chunk="dashboard-07-chunk-0">
-                <CardHeader>
-                  <CardTitle>Product Information</CardTitle>
-                  <CardDescription>
-                    Please provide the product name and description.
-                  </CardDescription>
-                </CardHeader>
-                <CardContent>
-                  <div className="grid gap-6">
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Product</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-
-              <Card x-chunk="dashboard-07-chunk-0">
-                <CardHeader>
-                  <CardTitle className="flex items-center justify-between">
-                    Delivery Eligibility
-                    <Switch />
-                  </CardTitle>
-                  <CardDescription>
-                    Please provide the product name and description.
-                  </CardDescription>
-                </CardHeader>
-                <CardContent>
-                  <div className="grid gap-6 grid-cols-3">
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Minimum Count</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Maximum Count</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Estimated Delivery Time</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-
-              <Card x-chunk="dashboard-07-chunk-0">
-                <CardHeader>
-                  <CardTitle className="flex items-center justify-between">
-                    Collection Eligibility
-                    <Switch />
-                  </CardTitle>
-                  <CardDescription>
-                    Please provide the product name and description.
-                  </CardDescription>
-                </CardHeader>
-                <CardContent>
-                  <div className="grid gap-6 grid-cols-3">
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Minimum Count</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Maximum Count</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Estimated Collection Time</FormLabel>
-                            <FormControl>
-                              <Input placeholder="Select product" {...field} />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            </div>
-
-            <div className="grid auto-rows-max items-start gap-4 lg:gap-8">
-              <Card x-chunk="dashboard-07-chunk-3">
-                <CardHeader>
-                  <CardTitle>Stock Information</CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <div className="grid gap-6">
-                    <div className="grid gap-3">
-                      <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                          <FormItem>
-                            <FormLabel>Stock Count</FormLabel>
-                            <FormControl>
-                              <Input
-                                placeholder="Enter product name"
-                                {...field}
-                              />
-                            </FormControl>
-                            <FormMessage />
-                          </FormItem>
-                        )}
-                      />
-                    </div>
-                  </div>
-                </CardContent>
-              </Card>
-            </div>
-          </div>
-        </form>
-      </Form>
+      <CreateInventoryForm  products={products}/>
     </ContentLayout>
   );
 }
